Add prompt for confirmation in install script if target file already exists

+++ install.sh
@@ -3,11 +3,28 @@
 SCRIPT_FILE="repox.sh"
 TARGET_LOCATION="/usr/local/bin/repox"
 
+overwrite=0
+
+# Warn the user if file already exists and ask for confirmation
+if [ -f "$TARGET_LOCATION" ]; then
+    echo "warning: $TARGET_LOCATION already exists."
+    while true; do
+        read -p "Would you like to overwrite the file? (y/n): " yn
+        case $yn in
+            [Yy]* ) overwrite=1; break ;;
+            [Nn]* ) exit ;;
+            * ) echo "Please answer y/Y or n/N.";;
+        esac
+    done
+fi
+
+# Do not prompt for confirmation if user agreed to overwrite
+if (( $overwrite == 0 )); then
+    read -p "This will copy ${SCRIPT_FILE} to ${TARGET_LOCATION}. Press [ENTER] to continue..."
+fi
+
 echo "Installing..."
-
-read -p "This will copy ${SCRIPT_FILE} to ${TARGET_LOCATION}. Press [ENTER] to continue..."
 cp $SCRIPT_FILE $TARGET_LOCATION
-
 
 if cp -rv $SCRIPT_FILE $TARGET_LOCATION; then
     printf "\n\tsuccess: file copied\n\n"
